Decompile freeFireslot

+++ src/include/game/game_097a50.h
@@ -139,7 +139,7 @@ u32 func0f0a5550(void);
 u32 func0f0a695c(void);
 u32 func0f0a6c30(void);
 u32 func0f0a70a8(void);
-s8 freeFireslot(s8 fireslot);
+s8 freeFireslot(s32 fireslot);
 u32 func0f0a70f8(void);
 u32 func0f0a7138(void);
 u32 func0f0a7d5c(void);
